+++ api/routes/course.routes.js
@@ -1,13 +1,6 @@
 const express = require("express");
 
 const { check } = require("express-validator");
-//const validateFields = require("../middleware/validateFields.middleware");
-const {
-  validateCourse,
-  validateFields,
-  validateCoupon,
-  validateDeleteCoupon,
-} = require("../middleware/validateFields.middleware2");
 
 const {
   allProjects,
@@ -22,18 +15,30 @@ const {
   oneCourse,
   allCourses,
 } = require("../controllers/course.controller");
+const {
+  validateCourse,
+} = require("../middleware/courseValidations.middleware");
+const {
+  validateCoupon,
+  validateDeleteCoupon,
+} = require("../middleware/couponValidations.middleware");
+const validateFields = require("../middleware/validateFields.middleware");
 
 const router = express.Router();
 
 //rutas de marcos
 router.get("/projects", allProjects);
+
 router.put(
   "/updateProject/:id",
   [check("id", "id is not type mongo").isMongoId()],
   updateProject
 );
+
 router.post("/addCoupon", validateCoupon, createCoupon);
+
 router.delete("/deleteCoupon/:id", validateDeleteCoupon, deleteCoupon);
+
 router.put(
   "/updateCoupon/:id",
   [check("id", "id is not type mongo").isMongoId()],
